pyproject.toml
[project]
name = "liqui-speak"
version = "0.1.0"
description = "One-command setup for real-time audio transcription with LFM2-Audio model"
readme = "README.md"
authors = [
    { name = "abhishekbhakat", email = "abhishek.bhakat@hotmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "click>=8.1.0",
    "requests>=2.31.0",
    "tqdm>=4.66.0",
    "pydub>=0.25.0",
    "soundfile>=0.12.1",
    "numpy>=1.24.0",
    "huggingface-hub>=0.20.0",
    "python-magic>=0.4.27",
    "pyaudio>=0.2.11",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.5.0",
    "pyright>=1.1.350",
    "flake8>=7.0.0",
    "flake8-import-order>=0.18.2",
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "black>=24.0.0",
]

[project.scripts]
liqui-speak = "liqui_speak.cli:main"

[tool.ruff]
target-version = "py312"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
    "E402",  # module level import not at top of file
    "TID252", # ban relative imports
    "W293", # blank line contains whitespace
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports

[tool.ruff.isort]
known-first-party = ["liqui_speak"]

[tool.pyright]
typeCheckingMode = "standard"
pythonVersion = "3.12"
include = ["src"]
exclude = ["**/__pycache__", "**/node_modules", "**/venv"]
reportMissingImports = "error"
reportMissingTypeStubs = "warning"
reportPrivateImportUsage = "error"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=liqui_speak --cov-report=term-missing"

[tool.black]
target-version = ["py312"]
line-length = 88
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[build-system]
requires = ["uv_build>=0.9.11,<0.10.0"]
build-backend = "uv_build"
